# docs/protocol/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Super STT daemon protocol",
    "description": "HTTP/1.1 + JSON over a Unix domain socket at `$XDG_RUNTIME_DIR/stt/super-stt-http.sock` (override with `SUPER_STT_HTTP_SOCKET`). There is no TCP listener: the socket's filesystem permissions are the first layer of access control, and the daemon reads `SO_PEERCRED` on each connection to identify the calling binary.\n\nEvery endpoint except `POST /v1/auth/request` requires `Authorization: Bearer <token>`. A token is minted only after the user approves your app in a consent popup, and is bound to the approved binary — an app cannot widen its own permissions. See `docs/protocol/auth.md`.\n\nBecause the transport is a Unix socket, the `servers` entry below is nominal; point your client at the socket and use any `Host`. With curl:\n\n```\ncurl --unix-socket \"$XDG_RUNTIME_DIR/stt/super-stt-http.sock\" \\\n     -H \"Authorization: Bearer $STT_TOKEN\" \\\n     http://stt.local/v1/ping\n```",
    "contact": {
      "name": "Super STT",
      "url": "https://github.com/jorge-menjivar/super-stt"
    },
    "license": {
      "name": "GPL-3.0-only",
      "identifier": "GPL-3.0-only"
    },
    "version": "0.2.4-beta.1"
  },
  "servers": [
    {
      "url": "http://stt.local",
      "description": "Nominal host; the transport is the Unix socket"
    }
  ],
  "paths": {
...
    "/v1/pipeline/{stage}": {
      "get": {
        "tags": [
          "pipeline"
        ],
        "summary": "Report one pipeline stage",
        "description": "The same object `GET /pipeline` carries in its array, for a client that only cares about one position. It is narrowed from that report rather than derived separately, so one stage and the whole list can never disagree.",
        "operationId": "get_stage",
        "parameters": [
          {
            "name": "stage",
            "in": "path",
            "description": "Pipeline position: `1` transcribes, `2` post-processes. A position that does not exist is a `404 unknown_stage`.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "The stage.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StageEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Token unknown, expired, or its binary changed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReasonEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks the `settings` scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No such stage (`unknown_stage`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Per-client rate limit hit; back off and retry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_token": [
              "settings"
            ]
          }
        ]
      },
      "post": {
        "tags": [
          "pipeline"
        ],
        "summary": "Select the backend filling a stage",
        "description": "Points a stage at an installed backend. Selecting a backend does not load a model — do that with `POST /pipeline/{stage}/model`.\n\nA backend that serves nothing this stage can run is refused: filling stage 2 with a transcription-only backend would leave the user staring at an empty model picker with no reason given.",
        "operationId": "set_stage_backend",
        "parameters": [
          {
            "name": "stage",
            "in": "path",
            "description": "Pipeline position: `1` transcribes, `2` post-processes. A position that does not exist is a `404 unknown_stage`.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            },
            "example": 1
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetBackendBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Selected.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StageMutation"
                }
              }
            }
          },
          "400": {
            "description": "No installed backend has that `source`, or it serves nothing this stage can run (`invalid_backend`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Token unknown, expired, or its binary changed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReasonEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks the `settings` scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No such stage (`unknown_stage`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Per-client rate limit hit; back off and retry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_token": [
              "settings"
            ]
          }
        ]
      },
      "delete": {
        "tags": [
          "pipeline"
        ],
        "summary": "Empty a stage",
        "description": "Deselects the stage's backend, unloading its model first if one is up. The stage then does nothing: stage 1 empty means no transcription, stage 2 empty means transcripts pass through unrewritten.",
        "operationId": "clear_stage_backend",
        "parameters": [
          {
            "name": "stage",
            "in": "path",
            "description": "Pipeline position: `1` transcribes, `2` post-processes. A position that does not exist is a `404 unknown_stage`.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Emptied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StageMutation"
                }
              }
            }
          },
          "401": {
            "description": "Token unknown, expired, or its binary changed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReasonEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks the `settings` scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No such stage (`unknown_stage`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Per-client rate limit hit; back off and retry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_token": [
              "settings"
            ]
          }
        ]
      }
    },
...
    "/v1/registry/backends/install": {
      "post": {
        "tags": [
          "registry"
        ],
        "summary": "Install a backend",
        "description": "Installs a backend from the registry, from a git repo, or from a local path — send exactly one of `source`, `repo_url`, or `local_path`.\n\nThe daemon picks the release asset matching this host's architecture and accelerators, and answers `202` as soon as that choice is made: the download itself continues in the background. Follow it on the `registry_install` event topic, keyed by the returned `install_id`. A `warning` means the install proceeded with a caveat worth showing the user.\n\nInstalling neither selects the backend nor loads a model — do that through `/pipeline/{stage}`.",
        "operationId": "install_registry_backend",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InstallRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted; the download runs in the background.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstallAccepted"
                }
              }
            }
          },
          "400": {
            "description": "Not exactly one of `source`, `repo_url`, `local_path` (`bad_request`), or no asset matches this host.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistryError"
                }
              }
            }
          },
          "401": {
            "description": "Token unknown, expired, or its binary changed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReasonEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks the `settings` scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No catalog entry for that `source` (`not_found`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistryError"
                }
              }
            }
          },
          "409": {
            "description": "An install for this backend is already in flight, or a recording is running (`backend_busy`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistryError"
                }
              }
            }
          },
          "429": {
            "description": "Per-client rate limit hit; back off and retry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_token": [
              "settings"
            ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorCode": {
        "type": "string",
        "enum": [
          "recording_in_progress",
          "download_in_progress",
          "no_switch_in_progress",
          "model_not_loaded",
          "invalid_model",
          "invalid_backend",
          "cuda_unavailable",
          "invalid_device",
          "invalid_audio_theme",
          "unsupported_language",
          "invalid_value",
          "not_found",
          "internal",
          "unknown"
        ]
      },
      "ErrorEnvelope": {
        "type": "object",
        "description": "The error envelope, as [`transport.md`] specifies it.\n\n`error_code` is the stable, machine-readable identifier clients switch on\nand the field that determines the HTTP status; `message` is prose for a\nhuman and may be reworded at any time.\n\n[`transport.md`]: https://github.com/jorge-menjivar/super-stt/blob/main/docs/protocol/transport.md",
        "required": [
          "status"
        ],
        "properties": {
          "error_code": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ErrorCode",
                "description": "The stable identifier for this failure. Absent only on an unclassified\nserver-side error."
              }
            ]
          },
          "message": {
            "type": [
              "string",
              "null"
            ],
            "description": "Human-readable detail."
          },
          "status": {
            "type": "string",
            "description": "Always `error`.",
            "example": "error"
          }
        }
      },
...
      "InstallAccepted": {
        "type": "object",
        "required": [
          "install_id",
          "source",
          "version",
          "selected_asset"
        ],
        "properties": {
          "install_id": {
            "type": "string"
          },
          "selected_asset": {
            "$ref": "#/components/schemas/SelectedAsset"
          },
          "source": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "warning": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "InstallRequest": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "source"
            ],
            "properties": {
              "source": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "repo_url"
            ],
            "properties": {
              "repo_url": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "local_path"
            ],
            "properties": {
              "local_path": {
                "type": "string"
              }
            }
          }
        ]
      },
...
      "Reason": {
        "type": "object",
        "description": "The `data` object of a [`ReasonEnvelope`].",
        "required": [
          "reason"
        ],
        "properties": {
          "reason": {
            "type": "string",
            "description": "Which case of `message` occurred — e.g. `expired`, `exe_changed`,\n`user_denied`."
          }
        }
      },
      "ReasonEnvelope": {
        "type": "object",
        "description": "The auth-failure envelope: `message` names the failure and `data.reason`\nsays which of its cases occurred.\n\nDistinct from [`ErrorEnvelope`] because the auth surface predates\n`error_code` and clients read `data.reason` there. Both are documented\nrather than reconciled, since changing either is a breaking wire change.",
        "required": [
          "status",
          "message",
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Reason"
          },
          "message": {
            "type": "string",
            "description": "The failure identifier, e.g. `invalid_session` or `auth_denied`."
          },
          "status": {
            "type": "string",
            "description": "Always `error`.",
            "example": "error"
          }
        }
      },
...
      "RegistryError": {
        "type": "object",
        "description": "The registry surface's error envelope, which carries one extra key.\n\nThese endpoints shipped before `error_code` existed, spelling the failure\nidentity as `error`. That key is still sent, because clients read it; the\nstandard `error_code` was added alongside rather than in place of it, so the\nwhole surface honors the \"`error_code` on every error\" rule without breaking\nanyone. Both name the same failure and always agree.\n\nDocumented as its own shape rather than folded into [`ErrorEnvelope`]:\n`error` appears *only* here, and putting it on the shared envelope would\ntell every other endpoint's reader to expect a key they will never receive.",
        "required": [
          "status",
          "error_code",
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "The same identifier under the key this surface has always used. Retained\nfor clients written against it; prefer `error_code`.",
            "example": "not_found"
          },
          "error_code": {
            "type": "string",
            "description": "The stable identifier for this failure.",
            "example": "not_found"
          },
          "message": {
            "type": [
              "string",
              "null"
            ],
            "description": "Human-readable detail, when there is any to add."
          },
          "status": {
            "type": "string",
            "description": "Always `error`.",
            "example": "error"
          }
        }
      },
...
      "SetBackendBody": {
        "type": "object",
        "description": "Which backend should fill the stage.",
        "required": [
          "source"
        ],
        "properties": {
          "source": {
            "type": "string",
            "description": "The backend's repo id, as `GET /backends` reports it.",
            "example": "github.com/acme/whisper"
          }
        }
      },
...
      "StageEnvelope": {
        "type": "object",
        "description": "One stage of the pipeline.",
        "required": [
          "status",
          "stage"
        ],
        "properties": {
          "stage": {
            "$ref": "#/components/schemas/StageReport"
          },
          "status": {
            "type": "string",
            "example": "success"
          }
        }
      },
      "StageMutation": {
        "type": "object",
        "description": "The answer to a stage mutation.\n\nThe two stages answer with different keys — stage 1 with `active_backend`,\nstage 2 with `post_processor` — because each grew its own endpoint before\nthe pipeline addressed stages by position. Both are documented rather than\nreconciled: changing either is a breaking wire change. Read the stage back\nwith `GET /pipeline/{stage}` for the one shape both share.",
        "required": [
          "status"
        ],
        "properties": {
          "active_backend": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ActiveBackend",
                "description": "Stage 1 only. `null` when the stage was emptied.\n\nDoubly optional for the same reason `resolved_accel` is: absent means\nthis was a stage 2 mutation, which reports `post_processor` instead;\n`null` means stage 1 itself is now empty."
              }
            ]
          },
          "message": {
            "type": [
              "string",
              "null"
            ]
          },
          "post_processor": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PostProcessorState",
                "description": "Stage 2 only."
              }
            ]
          },
          "status": {
            "type": "string",
            "example": "success"
          }
        }
      },
...
    },
    "securitySchemes": {
      "session_token": {
        "type": "http",
        "scheme": "bearer",
        "description": "Session token from `POST /v1/auth/request`. Bound to the calling binary and valid for 30 days."
      }
    }
  },
...
}